+++ omopcdmlib/Models/CohortAttribute.cs
@@ -5,13 +5,13 @@ namespace omopcdmlib.Models
 {
     public partial class CohortAttribute
     {
-        public long Id { get; set; }
-        public long CohortDefinitionId { get; set; }
+        public int Id { get; set; }
+        public int CohortDefinitionId { get; set; }
         public byte[] CohortStartDate { get; set; }
         public byte[] CohortEndDate { get; set; }
-        public long SubjectId { get; set; }
-        public long AttributeDefinitionId { get; set; }
+        public int SubjectId { get; set; }
+        public int AttributeDefinitionId { get; set; }
         public byte[] ValueAsNumber { get; set; }
-        public long? ValueAsConceptId { get; set; }
+        public int? ValueAsConceptId { get; set; }
     }
 }
