+++ omopcdmlib/Models/ConceptRelationship.cs
@@ -5,9 +5,9 @@ namespace omopcdmlib.Models
 {
     public partial class ConceptRelationship
     {
-        public long Id { get; set; }
-        public long ConceptId1 { get; set; }
-        public long ConceptId2 { get; set; }
+        public int Id { get; set; }
+        public int ConceptId1 { get; set; }
+        public int ConceptId2 { get; set; }
         public string RelationshipId { get; set; }
         public byte[] ValidStartDate { get; set; }
         public byte[] ValidEndDate { get; set; }
