+++ omopcdmlib/Models/ConceptSynonym.cs
@@ -5,9 +5,9 @@ namespace omopcdmlib.Models
 {
     public partial class ConceptSynonym
     {
-        public long Id { get; set; }
-        public long ConceptId { get; set; }
+        public int Id { get; set; }
+        public int ConceptId { get; set; }
         public string ConceptSynonymName { get; set; }
-        public long LanguageConceptId { get; set; }
+        public int LanguageConceptId { get; set; }
     }
 }
