+++ omopcdmlib/Models/DoseEra.cs
@@ -5,11 +5,11 @@ namespace omopcdmlib.Models
 {
     public partial class DoseEra
     {
-        public long Id { get; set; }
-        public long DoseEraId { get; set; }
-        public long PersonId { get; set; }
-        public long DrugConceptId { get; set; }
-        public long UnitConceptId { get; set; }
+        public int Id { get; set; }
+        public int DoseEraId { get; set; }
+        public int PersonId { get; set; }
+        public int DrugConceptId { get; set; }
+        public int UnitConceptId { get; set; }
         public byte[] DoseValue { get; set; }
         public byte[] DoseEraStartDate { get; set; }
         public byte[] DoseEraEndDate { get; set; }
