+++ omopcdmlib/Models/DrugStrength.cs
@@ -5,16 +5,16 @@ namespace omopcdmlib.Models
 {
     public partial class DrugStrength
     {
-        public long Id { get; set; }
-        public long DrugConceptId { get; set; }
-        public long IngredientConceptId { get; set; }
+        public int Id { get; set; }
+        public int DrugConceptId { get; set; }
+        public int IngredientConceptId { get; set; }
         public byte[] AmountValue { get; set; }
-        public long? AmountUnitConceptId { get; set; }
+        public int? AmountUnitConceptId { get; set; }
         public byte[] NumeratorValue { get; set; }
-        public long? NumeratorUnitConceptId { get; set; }
+        public int? NumeratorUnitConceptId { get; set; }
         public byte[] DenominatorValue { get; set; }
-        public long? DenominatorUnitConceptId { get; set; }
-        public long? BoxSize { get; set; }
+        public int? DenominatorUnitConceptId { get; set; }
+        public int? BoxSize { get; set; }
         public byte[] ValidStartDate { get; set; }
         public byte[] ValidEndDate { get; set; }
         public string InvalidReason { get; set; }
