+++ omopcdmlib/Models/ProcedureCost.cs
@@ -5,10 +5,10 @@ namespace omopcdmlib.Models
 {
     public partial class ProcedureCost
     {
-        public long Id { get; set; }
-        public long ProcedureCostId { get; set; }
-        public long ProcedureOccurrenceId { get; set; }
-        public long? CurrencyConceptId { get; set; }
+        public int Id { get; set; }
+        public int ProcedureCostId { get; set; }
+        public int ProcedureOccurrenceId { get; set; }
+        public int? CurrencyConceptId { get; set; }
         public byte[] PaidCopay { get; set; }
         public byte[] PaidCoinsurance { get; set; }
         public byte[] PaidTowardDeductible { get; set; }
@@ -16,8 +16,8 @@ namespace omopcdmlib.Models
         public byte[] PaidByCoordinationBenefits { get; set; }
         public byte[] TotalOutOfPocket { get; set; }
         public byte[] TotalPaid { get; set; }
-        public long? RevenueCodeConceptId { get; set; }
-        public long? PayerPlanPeriodId { get; set; }
+        public int? RevenueCodeConceptId { get; set; }
+        public int? PayerPlanPeriodId { get; set; }
         public string RevenueCodeSourceValue { get; set; }
     }
 }
